refactor(#2456): split grip_modelling_tab.py into widgets/coordinator modules
- Extract PressureVisualizationWidget + ContactMetricsWidget to
  _grip_modelling_widgets.py (213 LOC)
- Slim grip_modelling_tab.py coordinator to ~884 LOC (was 1066 LOC);
  retains GripModellingTab + re-exports
- All public names re-exported for backward compatibility
- TDD contract tests added (3 passed, 3 skipped pending mujoco install)
